+++ src/app/product/services/product-service.service.ts
@@ -57,7 +57,10 @@ export class ProductServiceService {
   getProductTotalRemaining(productId: string): Observable<number> {
     return new Observable((obs: Subscriber<number>) =>
       this.getProductsInStock(productId).subscribe(
-        pis => obs.next(pis.reduce((sum, val) => sum += val.Remaining, 0))
+        pis => {
+          obs.next(pis.reduce((sum, val) => sum += val.Remaining, 0));
+          obs.complete();
+        }
       ));
   }
 
@@ -65,7 +68,10 @@ export class ProductServiceService {
     if (productId) {
       return new Observable((o: Subscriber<ProductInStock[]>) => {
         this.httpEmulator.subscribe(productInStock =>
-          o.next(productInStock.filter(pis => pis.Product.Id === productId)));
+          {
+            o.next(productInStock.filter(pis => pis.Product.Id === productId));
+            o.complete();
+          });
       });
     } else {
       return this.httpEmulator;
@@ -83,6 +89,7 @@ export class ProductServiceService {
         obs.next(
           productsInStock.find(pis => pis.Product.Id === productId)?.Product
         );
+        obs.complete();
       });
     });
   }